Move types UserData and CachedUser back to LocalStorageKeystore
