+++ src/api/index.ts
@@ -2,8 +2,8 @@ import axios, { AxiosResponse } from 'axios';
 import { Err, Ok, Result } from 'ts-results';
 
 import { jsonParseTaggedBinary, jsonStringifyTaggedBinary, toBase64Url } from '../util';
-import { CachedUser, makePrfExtensionInputs } from '../services/keystore';
-import { LocalStorageKeystore } from '../services/LocalStorageKeystore';
+import { makePrfExtensionInputs } from '../services/keystore';
+import { CachedUser, LocalStorageKeystore } from '../services/LocalStorageKeystore';
 import { UserData, Verifier } from './types';
 import { useEffect, useMemo } from 'react';
 import { UseStorageHandle, useClearStorages, useSessionStorage } from '../components/useStorage';
